+++ src/_main_/celery/app.py
@@ -2,11 +2,11 @@ from __future__ import absolute_import, unicode_literals
 import os
 from celery import Celery
 from celery import shared_task
-from . import celeryconfig
+from _main_.celery.config import CeleryConfig
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "_main_.settings")
 app = Celery('massenergize_celeryapp')
-celery_config = celeryconfig.CeleryConfig().get_config()
+celery_config = CeleryConfig().get_config()
 
 app.config_from_object(celery_config)
 app.autodiscover_tasks()
